backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
alembic==1.13.1
httpx==0.26.0
pyjwt==2.8.0
cryptography==42.0.2
redis==5.0.1
langchain==0.1.5
langchain-google-genai==0.0.8
langchain-core==0.1.18
pandas==2.2.0
numpy==1.26.3
gymnasium==0.29.1
shimmy>=1.3.0
stable-baselines3==2.2.1
torch==2.1.2